refactor: 회원 서비스 외부 및 내부 API 분리
